Improve experimental data reporting.
