+++ benchmarks/definitions.py
@@ -16,7 +16,7 @@ kw_steps = "#st"
 kw_tlast = "Tlast"
 kw_ver = "ver.%"
 kw_toolname = "Synduce"
-kw_tool_main_algo = "S-CEGIS+PB"
+kw_tool_main_algo = "S-CEGIS${\\bf +PB}$"
 kw_tool_main_algo_v2 = "SE²GIS"
 kw_baseline = "Baseline"
 kw_segis = "Symbolic CEGIS"
